Rewrite user command processing in lldb style
Instead of always providing a context menu and a list of commands
the user can execute - allow user to execute any possible command
at any time. It makes it less beginner friendly, since instead
of having a list of all the possible commands always displayed to
you in every situation you are now expected to remember all the
possible commands. I'm writing this for myself though, so it's
okay i'll handle that.