Slightly improve the naming convention
For cases when a variable name conflicts with it's type name -
change variable name: make it more specific to the context in
which it is being used.
Task index/id is being used interchangeably in code, thus - always
use id, since it is shorter and makes more sence in some cases.