Rework the way file links are displayed to the user
Instead of a displaying a separate list of just links to files,
transform stdout/stderr of executing tasks on the fly and
highlight those links with a different color right in there.
This way it is much easier for the user to figure out which
link to open.
Also, when user enters 'o' with invalid index - display the
whole combined output with highlighted links again.