Refactor the way task output is being postprocessed/printed
Right now as soon as task produces an output - that output gets
added to the task_output struct, which then get's processed with
find_and_highlight_file_links(). This is fine right now when we are
always interested in file links of the entire task output.
This is however not the case with the future gtest integration.
1) We don't want to automatically print all the test execution output.
We want to buffer it instead and only show portions of it on demand
(failed tests for example).
2) We don't want to scan entire gtest task output for file links.
This is not a huge deal, but I would want the found links in the
output of a single test start from index '1' instead of '112'.

With this change, task_output now only stores output_lines that
are being printed to the console and find_and_highlight_file_links()
only scans the lines that are being printed to the console.
In case task's output is being streamed to the console in real time,
output_lines will be constantly moved from task_execution to
task_output.
In the opposite case, output_lines are kept in task_execution until
the latter one finishes.