Implement "__gtest" tasks
When task with "command" "__gtest <gtest-binary-path>" is run
parse output of the task as a gtest binary output.
If the task is a pre_task - only display test results on failure.
If the task is a primary task - display test execution progress
and test results.
If all the tests have completed successfully - only display one line
saying it plus some statistics.
If there are failed tests - display only failed tests plus some
statistics.