Try to fix a rare "bus error" when running a __gtest task
This is a blind attempt to fix the problem.
What might be happening:
parse_gtest_output() expects to receive messages from gtest
in following order:
[===][RUN][FAILED][RUN][FAILED][===]
In reality however it gets following:
[===][RUN][FAILED][RUN][FAILED][===][FAILED][FAILED]
Vast majority of the time we receive last [===] gtest line
together with the rest of the [FAILED] in a single
task_execution.output_lines batch. We get to the [===] a break
the loop. Thus we ignore those additional [FAILED] lines
right then and there.
What might be happening sometimes is that we don't receive
all [FAILED] messages in the last batch. Some of them
we receive later. In such case we get into a [FAILED] "else if"
that expects gtest_exec.current_test to be set, but it is not
and thus it has a garbage in it. We then potentially use that
garbage as a pointer in memory, which might be leading to these
"bus error"s or "segmentation faults".
With current fix parse_gtest_output() will stop parsing
any task output after parsing the last [===].