Implement "g<ind>" command
Implement command, that allows you to view output of a google test.
In case there are failed tests "g" allows to view list of failed tests
and view output of any of them.
In case there are no failed tests "g" allows to view list of all
tests and view output of any of them.