Fix sporadically failing tests due to signal handling
notify_task_execution_about_child_process_exit() signal handler
used to call enqueue(), which is not async-signal-safe. This led
to malloc() being called "recursively" sometimes (when thread calls
it and in the middle of it the signal handler triggers and calls it
again). On Mac it causes "Trying to recursively lock an os_unfair_lock".
Get rid of SIGCHLD handler. Instead - make a patch
for tiny-process-library, that allows specifying a callback, that is
called when a process finishes. "git apply" that patch on cmake initialization
if it is not already applied. Also fail cmake if git is not installed.
Revert previous fix for a race condition between "exit" task and
last "stdout/stderr" tasks because now that race condition can't
happen anymore.