Fail __gtest task if specified binary is not a gtest executable
If I create a __gtest pre_task with incorrect binary, the task
would just always complete successfully, which would make me
think that the tests have succeeded (when in fact they have not).
If __gtest binary is not a gtest binary - notify user about it
and fail current task_execution.