Don't re-execute pre_tasks on "tr"
I plan to use "tr" to re-execute tests to troubleshoot
sporadically failing tests. In such situation re-executing
all of the pre_tasks (which are typically "build tests") does not
make any sense.
There could potentially be use-cases to re-execute pre_tasks
together with the primary task, but nothing comes to my mind
right now. If there would be a necessity in this - I will
re-implement "tr". For now, thouh, it is just fine the way
it is.