Rewrite cdt data structures in entity-component-system fashion
It expands possibilities in general but also allows us to
have multiple gtest_execution instances at the same time, while
also having some of them for executions, that will be run in
the future. Currently if we have any number of gtest_executions
all of them can only be related to the current/next execution.
With the new approach we can create gtest_executions that are
related to executions, that will be executed much later.