Implement "gs" command
Implement command, that allows searching through a google test
output with a regular expression.
Fix typo in "s" command description.
Fix a bug in search_through_text_buffer() displaying invalid
indexes of found lines when working with a subset of the
whole buffer (in case of gtest buffer for example).