Resolve #14: Simplify tests
Code duplication: my thoughts right now are that expected
data duplication in tests is fine. It is easier to understand
a test when it has actual expected data declared right in the
test body as opposed to being declared somewhere in the common
code in some of the macros (no idea in which exactly). Writing
tests in such a way is faster since you don't need to constantly
extract common code and look for a common code to reuse.
In an environment with multiple developers such constant meddling
with common code would lead to frequent merge conflicts.
It only makes sence to extract a very well-known expected data
into constants.
Sure, tests become larger and making a global change in expected
data is more difficult. If such global change would actually
happen and irritate me enough - I might reduce expected data
duplication.