Resolve #1: Implement task config profiles
Each profile is just a bunch of variables with values.
Tasks can have placeholders in their attributes for those variables, so that
when config is being loaded - those placeholders are replaced with values
of the corresponding variables of the currently selected profile.
If no profile is explicitly selected via command line arguments - the first
profile from the list is selected by default.