Fix possible race condition in the Ctrl+C handler
Executing StopRunnningProcessesOrExit() inside of the Unix
implementation of the SetCtrlCHandler() is fine, since it
will get called from the same thread from a signal handler.
Doing so with Windows implementation can cause a race, because
Ctrl+C handler gets called from a different thread.
We also can't just put it behind a mutex, because we are
not allowed to acquire mutex from Unix's SetCtrlCHandler() since
it is a signal handler.
Solution: make the whole Ctrl+C handler OS-specific.
Also in Windows don't put the whole entt::registry behind mutex.
Instead - keep a side vector of active PIDs and use it
in the Ctrl+C handler (always serialize access to that vector).