Fix the way tests validate FinishProcess() calls
Due to how OsApiMock::proc_exit_info is implemented, the same
shell_command can be started multiple times and FinishProcess()
will be called only once. While tests will consider this correct,
it is not and it leads to active_process_ids in os-api-*.cpp not
being properly cleared and as a result - Ctrl+C no longer works
after running "tr" or "gtr".
Make OsApiMock::proc_info use PID as a key instead, so that the
aforementioned case would lead to test failures.
Currently all "tr"/"gtr" tests fail (which is expected for now).