Replace EXPECT_PROCESS() with EXPECT_PROCS()
Previous macro only allowed to check that the specified process
has been started and finished N times.
New macro allows to check that a sequence of processes
has been started and finished.
The sequence does not have to be exact: there can be
process calls in between the processes specified in sequence.
In case specified sequence of process calls can't be found,
macro will print the expected sequence of processes and
the actual sequence of ALL executed processes, while also
displaying which of the executed processes have not been finished
if any.
In case specified sequence of processes is found but not all
of them finished, macro will print list of processes, while
displaying which of them have not been finished.