Read STDIN in UTF8 on windows
On Windows there are at least two configurations for stdin:
1. when running from within mintty
2. when running from Command Prompt or Microsoft Terminal
In case 1. the code is exactly the same as on Mac, because
mintty already provides STDIN in UTF8.
In case 2. the only thing I was able to to achieve is to
read STDIN in UTF16 (wstring) and then convert it to UTF8
manually.
Doing the necessary setup calls from 2. in case 1. breaks
mintty's input: it still remains sort of UTF8, but adjacent
bytes are switched: if you type привет it will give you
рпвите (and that's in a good case, in reality when you will
try to print it - it will print garbage).
The best way I've found how to differentiate between 1. and 2.
is checking if "TERM" ENV variable is set (it is set in 1.
and is not set in 2.). Although it is not ideal and may change in
the future (hell, even now if you start Command Prompt from mintty
"TERM" variable will be set). It would be great to find another
way to differentiate.