Add googletest as a dependency
I was thinking long and hard whether to have tests for this
project or not.
An undeniable pro of having tests:
When you change existing code you might not remember all the test
cases that you need to execute manually. In automated tests
you store knowledge about all of those use-cases. And all of them
are getting executed really fast and without much effort.
There are difficulties in writing tests, though they could be
resolved:
- don't use common expected data among tests (it requires you to
know all of it and how it interacts with each other upfront),
instead have all tests have their own local expected data but
provide functions to simplify generation of that data
- in SetUp() initialize only that stuff that is used in every
test and should not be different across different tests (all the
rest of the stuff should be initialized in the tests themselves
and the initialization can be simplified by having good
initialization functions for all kinds of scenarios)
- don't set up mocks for async calls, instead - register callbacks
specified in those calls and trigger those callbacks with expected
data right in the middle of the test execution