Implemenet means of handling callbacks inside Processes
Implement ProcessRuntime::SharedPtr() that returns THE shared
pointer to the specified process directly from the processes
table.
What this enables Processes to do is to call async functions, that
accept callbacks, and pass that shared pointer inside of those
callbacks. In cases when a Process gets cancelled early and its
entry gets removed from the processes table - the shared pointer
keeping the Process object alive would still live, because it
was passed by copy inside that callback function. In such case
when that callback function would get called - it won't fail
because the shared pointer will still hold a valid Process object.