Implement background IO multithreading support
Threads::ScheduleIO() will be used to execute the on_background()
callback on the IO thread, return value of which will be passed
as an argument to the on_main() callback that will be executed
on the UI thread.
To keep it simple, value is passed between the threads by value,
thus passing large objects is better to be done via pointers.
ScheduleIO() will not catch exceptions thrown by on_backoground().
The latter one is expected to handle them on its own.