Separate Process source files from the rest of source files
In terms of configuration there needs to be some processes as well
as some structs. The two can't be in the same source file because
Application depends on those structs while processes depend on the
Application.
Having a source file per each class/struct would've solved that
problem but for now lets keep source files with processes separate
inside the process/ directory (so that we can have a config.hpp
with structs and process/config.hpp with processes
at the same time).