Connect InputAndListView.qml with its C++ counterpart
Replace hardcoded text values with values from C++.
Create callbacks in C++ for all the necessary user events
that would need to be handled. Make clicking on a list
item trigger the same callback as pressing enter on the
input field.
Allow disabling the button from C++: in terms of opening
a project the button should not be active until the value points
to an actual file that can be opened.
List model is not implemented yet in C++.