Improve debugability of ProcessRuntime
Make IsAlive() require only ProcessId so that it can be called
in contextes where the entire Process is not available.
Make SharedPtr() crash if specified process not IsAlive().
Make Schedule() crash if specified parent process not IsAlive().
Implement PrintProcesses() that prints process table. The new
method is intended to be used from inside the debugger.
Get rid of kNoopExecute constant. Instead, introduce a Noop()
base method. Also add a KeepAlive() method. It is intended to be
used when Process should not finish when all of its child
processes finish.