Simplify management of child processes in ProcessRuntime
Replace Process.child_ids (that stores IDs of both active and
finished child processes) with Process.running_child_ids (that
stores IDs of active child processes).
Get rid of ProcessRuntime.finished.
The idea of storing children in process table while their parent
process is active was based on the assumption that the process
table would just be a Vector<Process> meaining that if you
delete a process from it - its memory gets de-allocated. Since
we are storing Vector<QPtr<Process>>, those interested in those
child processes typically just obtain a copy of that QPtr and
store it locally so even if we remove a process from the table -
all the interested clients would still be able to retain their
child process instances.