Slightly simplify callbacks of InputAndListView
Instead of having one callback for selecting and item from the
list and another callback for actually confirming selection, have
only one callback that confirms selection and provides the index
of the selected item.
This kind of behavior will be common across different views, that
will have a list of items and a context menu. There each item
of the context menu will have a callback, that will have an index
of the item of the list passed to it as an argument.