De-generalize InputAndListView into OpenProjectView
It is much easier to generalize a bunch of use-case-specific
views than the other way around. Thus from now on, all views
will be created specifically for their use and only AFTER
they are found to be useful in multiple uses - they would get
generalized.
Also don't add methods for each individual view to UserInterface.
It should only have methods for views, that are used in MULTIPLE
different use-cases and called from different call sites in
the code base. Views specific to one place in the code, should
be initialized in that place in the code. To make data/list field
names easier to manage in such case (to reduce typing) introduce
new naming schema:
- v<variable name> - variable that belongs to the viewSlot
- d<variable name> - variable that belongs to the dialogSlot
- s<variable name> - variable that belongs to the statusSlot