Fix Process class name resolution in MSVC
In MSVC __FUNCTION__ used inside the constructor
produces a long function name: <class>::<method>.
In our case we want to determine the class name
inside the constructor so we need either of those
but not both.
Use __func__ keyword to do that instead. It works
consistently both in MSVC and LLVM.