Implement named processes
Currently if we start a process that interacts with the current
view and then start another process - the view will get changed
properly as well as the event listeners but the old process itself
will hang there forever.
This change aims to handle that case. Each process that interacts
with the UI gets assigned a name (typically a UI slot name).
This way when we for example start a SearchUserCommands process,
we cancel the current active process, that uses the dialogSlot,
but we don't cancel the process that is assigned to the viewSlot.
But in case we start a process that will change the current view -
we will cancel all the named processes and start that new process
with viewSlot name.