Fix crash on concurrent finish of a child and its parent
When a child cancels its parent and finishes itself at the same
time, and in case parent gets picked up from the to_finish set
first - while finishing the parent we will also finish the child.
But then we will continue to iterate over to_finish and try to
finish the child again, although it is no longer there in the
processes and we either will finish a wrong process the took that
spot or crash because the spot contains NULL.
When iterating over to_finish - first check if the specified
process exists and has the same id and only after that proceed
to finish it.