Fix premature dequeue of an event in some cases
There might be a case when we start executing processes without
any events enqueued, but by the time we end - some of
the processes has already enqueued one new event. Currently we
would just dequeue it assuming that it was there before we've
started and that we've already processed it. That will make us
skip that event altogether.
Instead - after executing processes, only dequeue an event that
was there before we've started.