Fix child process callback handling on windows
Currently when we fail to start a process in ExecOSCmd, we receive
the errorOccurred() signal from QProcess. Inside of it we call
WakeUpAndExecuteProcess(). On Mac it works fine because
the callback gets called in a different event loop iteration
relative to the iteration where we've started the process.
On Windows however the callback gets called from inside the same
stack. This leads to ExecuteProcesses() being called from inside
ExecuteProcesses(). This poses a problem because in case we have
an "execTask" event active the inner calls to ExecuteProcesses()
keep on "handling" that event and keep on spawning new ExecOSCmd,
because the outermost ExecuteProcesses() haven't reached the part
where it dequeues that event yet.
To fix this - make sure QProcess callbacks are always handled in
a different event loop iteration to prevent the aforementioned
function from being called inside of itself.