Allow selecting executions in the execution history UI
When an execution is manually selected by the user, in case
a new execution gets created - the selection won't automatically
switch to the new one while the manually selected one is present.
Only if user selects either the latest running execution or the
latest execution overall - then the selection will get reset and
will start to following the latest execution automatically.