Get button brightness under control
When button is getting pressed - don't make it even lighter than
when hovered. Instead - make it go back to its normal brightness.
This way it feels like when you hover over it it gets elevated
(becoming brighter due to receiving more light) and when you click
on it - it gets pressed down (becoming darker again).