Make app not crash when closed too early
When you close an app while InformationProvider's background
thread task is still running - the app crashes when invokeMethod()
is called with a pointer to InformationProvider, that has been
already deleted on the main thread by that time.
Instead - pass QCoreApplication pointer as context into
invokeMethod(). After the app is closed - that pointer will
be null and invokeMethod will not do anything when receiving
a null context.