Improve efficiency of QVariantListModel
Currently every list stores a 3 copies of itself:
1) list of original data
2) unfiltered list of QVariantList objects
3) filtered list of QVariantList objects
I know that memory is not a problem and stuff but the is not
efficient at all (especially for larger lists).
Get rid of unfiltered list in QVariantListModel and make it
compute filtered list based on the original list on the fly.