Fix movement in RunTask
When you open RunTask and it loads a list of tasks, if you open
SearchUserCommandDialog - it will capture the focus, but at the
same time in the background RunTask will finish loading tasks
and will steal the focus from the dialog. This should not happen,
dialog should retain the focus regardless of what happens in
the background.
Rework RunTask so that it acquires focus only the beginning of its
creation. To do that - don't use Loaders.
As a nice bonus - have a placeholder (displayed while loading
tasks) as a part of SearchableTextList.