Refactor TaskExecutor API
Support async loading of TaskExecutions and their outputs
as a separate methods. The separation is due to outputs possibly
being too large (hunders of kilobytes or even megabytes in size)
to allocate in memory for ALL executions being fetched. Maybe
not too large but definitely wastefull and time consuming (memory
allocation).