Get back to using KeyNavigation
Previously moving focus from TextArea to the root component
of ReadOnlyTextArea was not possible due to ScrollView:
if a child component with a focus does not accept an event -
the event gets sent to its parent component. This process
continues until a component in the chain accepts the event.
This didn't work that way for ReadOnlyTextArea, because TextArea
is inside ScrollView and the later always accepts arrow key
events.
Use Keys.forwardTo to redirect events, reaching ScrollView from
TextArea, to the root FocusScope. This way the KeyNavigation,
assigned on to the FocusScope from the client code, will receive
those key events.