Fix SearchableTextList sorting of filtered results
When sorting we compare all searchable_roles of all rows of the
list. In some cases the filter term might not match the first
searchable role at all but match the following role. In such
situation GetSortableProps() will get called with a string, that
does not have any highlights in it and return SortableProps with
first_match_pos remaining -1. This will lead to the row being
always first in the list (because of the smallest first_match_pos).
Fix GetSortableProps() to return SortableProps with max possible
values so that they always end up in the end of the list instead.
And when two rows with first searchable_role not matching will
get compared - they will be properly sorted by their second or
third searchable_role.