Fix background highlight of TextList item on right mouse click
Currently when you right click an item to open context menu -
the background of an item becomes transparent instead of light
grey.
Make the color condition differentiate between left/right mouse
buttons being pressed, because it should only have transparent
background in the moment when a left button is pressed.