Make use of UiCommandBuffer for inserts and removals as well
Inserts and removals might not really need it that much because
they are fast enough even on Windows in debug build. However there
is a case where when you have an empty list and you insert
all the items into it - the interface can lag for a moment (in
release mode it does not have this lag). With this change in place
that lag is either much less noticeable or absent entirely.
The change however future-proofs the list, because currently I've
only tested 50~ elements lists and while on such list sizes
distribution of inserts and removes does not make much difference,
it might on larger lists.