Drastically improve searching for documentation
Set documentation file name as a title and absolute path as
subtitle. This way looking for "qstring" or "basic_string" is
easier since the file name itself matches the search term, but
in case you want to look for a documentation in a specific
folder - you can do that with the subtitle search.