Fix regression in ChooseFile caused by fuzzy-search overhaul
Make minimal sub-match length configurable, since in ChooseFile
it needs to be 1 instead of the default 2.
Also when reloading the list of suggestions in ChooseFile, in
case SetFilter() didn't call Load() - call it ourselves. This
happens when we've just refreshed the suggestion list from disk
and file (and thus filter) is still empty.