Fix ApplicationWindow's maximized/normal status after reset
When a window is maximized and user resets window's size to the
default one - window's maximized/normal show status does not
change accordingly (because window is only maximized/normalized
when initially displayed). As a result when user starts the app
the next time - window is still displayed maximized.
Dynamically detect changes to viewSystem.dimensions.isMaximized
and apply them to window during runtime so resets of dimiensions
is properly applied to the window and saved to database.