Fix PageUp/PageDown in TextArea in some cases
1) When you open a fresh TextArea with empty cursor history -
PageUp/PageDown does not work due to cursor history being empty.
Consider current cursor position to be 0 in such case.
2) When you move cursor to the absolute bottom of TextArea, it
gets outside the bounds of actual QString, stored in TextArea.
Doing PageDown in such case causes SEGFAULT due to trying to
access a char of QString out of bounds. In case current cursor
position is outside of QString - make it point to the last char
of the string.