Improve last fix of TextArea's upward navigation
If you open search bar and close it - after that the upward
navigation breaks again because textArea still has KeyNavigation
of searchBar attached to itself (because setting
KeyNavigation.down: null on searchBar does not reset the "up"
navigation implicitly attached to textArea).
This fix makes it so when searchBar gets hidden - "up" navigation
of textArea gets reset appropriately as well.