Fix opening terminal on Mac working sporadically
Path to folder to cd in was being passed as an env variable
right inside the shell, that will be opened by terminal, which
will be opened by osascript. This used to work when osascript
was opening the first instance of terminal (while no other
instance is running). It does not work however when another
terminal is already running. In such case osa won't be forking
a terminal process child and that environment variable won't
be delivered.
Pass the folder via command line arguments of osa instead.