Make git diff fit the GitCommit screen
You can't really do a multiline select with such
approach, but the desired UI with a multiline
select possible can't be implemented with OOB
QML's components. A completely custom component
(like a custom TextArea) would need to be
implemented (to be used for a side-by-side diff
where each side is a separate TextArea) which
would display the same number of lines taking
into account line wrapping.