Somewhat fix typing special chars in readonly TextArea
Some key combinations (like Alt+N or Alt+U on
Mac) bypass Keys.onPressed of TextArea (meaning
that the said signal does not get triggered for
them) and as a result end up being typed in a
readonly TextArea.
Using "readOnly" of OOB TextArea is not an option
because:
a) it make cursor invisible and there is nothing
you can really do about it
b) it also disables such important shortcuts as
Ctrl+F
The only possible solution right now is to re-set
OOB TextArea's text to Cdt.TextArea.text when
its onTextChanged gets called while readonly is
true and when both texts are not equal (such
inequaity means that the text was not changed
via Cdt.TextArea's API but by Qt's internal
key press handling.