Make new TextArea use new TextFormatter class
QTextBlock::firstLineNumber() no longer provides
accurate result (since it is a part of a 1-line
EditText from which the whole TextArea is
composed). Thus passing it as an argument is no
longer useful. And since we need to change such
a commonly used signature lets also fix some
other errors such as introduce a TextFormat
struct instead of TextSectionFormat.