Fix qDebug not being picked up by QProcess
On Windows when Qt GUI app A runs a Qt app B, the qDebug() calls made
by B won't be visible to QProcess of A. At the same time std::cout and
std::cerr calls of B will be visible to A.
This seems to be caused by the fact that since A doesn't have
a Windows Console of its own - B does not get any console either.
This does not affect std's functionality because (according to ChatGPT)
for std Windows creates an implicit hidden Console. That console
however does not get used by qDebug() for some reason.
To solve this issue - introduce a settings option (that is only
available on Windows) that, when enabled, makes TaskSystem create
processes for its tasks using CREATE_NEW_CONSOLE. This however creates
an ugly window for the said console.
This functionality is optional because:
a) the ugly window
b) its only useful with Qt apps