Add support for half-staged files in GitCommit
Previously if you were to make changes to a file, stage them and then
make more changes to the file - it would be displayed as completely
staged while in fact only the first batch of changes are staged and
the last batch is still only in the working tree.
Change status icon for unstaged and staged files in GitCommit while
also introducing a new separate icon for half-staged files.
When toggling half-staged file - toggle it to a completely staged
state.