Fix fuzzy search missing 2 sub-match results
Currently when searching for something like "libtorrent-docs/session"
using search term "torrentsession" we successfully find "torrent" but
fail to find the "session" part. This is because while going through
"docs" we stumble upon the "s" there and start matching the "session"
against it. We fail and instead of rolling back 'ti' to position of
"s" in the search term we instead roll it back to "t" right before it.
As a result when we next stumble upon "s" in "/session" we fail to
start the match because the 'ti' index is pointing at "t" instead of
"s".